@@ -1,9 +1,10 @@
 'use strict';
 
-var React = require('react');
-var Facebook = require('./facebook');
+import React from 'react';
+import ReactDOM from 'react-dom';
+import Facebook from './facebook';
 
-React.render(
+ReactDOM.render(
   <Facebook
     title="Welcome to the Mullet Stack."
     subtitle="Facebook in the front. Walmart in the back."
