@@ -2,21 +2,22 @@ var path = require('path');
 var webpack = require('webpack');
  
 module.exports = {
-  entry: './src/react_components/main.js',
+  entry: [
+//    'webpack-hot-middleware/client?path=/__webpack_hmr&timeout=20000',
+    './src/react_components/main.js'
+  ],
   output: {
     path: path.join(__dirname, 'public/js'),
     filename: 'app.built.js'
   },
   module: {
     loaders: [
-      {
-        test: /.jsx?$/,
-        loader: 'babel-loader',
-        exclude: /node_modules/,
-        query: {
-          presets: ['es2015', 'react']
-        }
-      }
+      { test: /.jsx?$/, loader: 'babel-loader', exclude: /node_modules/, query: { presets: ['es2015', 'react'] } }
     ]
-  },
+  }
+//  plugins: [
+//      new webpack.optimize.OccurenceOrderPlugin(),
+//      new webpack.HotModuleReplacementPlugin(),
+//      new webpack.NoErrorsPlugin()
+//  ]
 };
